Fix app
Now actually does something useful.

Note that there's a slightly awkward interaction between this package and `wai-app-static`. A `*.x.br` is only served if `*.x` is present on disk, even though the contents of the latter don't matter. This is because `wai-app-static` does it's own existence check, and generates a builder response rather than a file response otherwise.